rename to ml-inference
