+++ data-prepper-plugins/ml-inference-processor/README.md
@@ -1,0 +1,39 @@
+
+# ml Processor
+
+This plugin enables you to send data from your Data Prepper pipeline directly to ml-commons for machine learning related activities.
+
+## Usage
+```aidl
+lambda-pipeline:
+...
+  processor:
+    - ml_inference:
+        host: "<your OpenSearch url>"
+        aws_sigv4: true
+        action_type: "batch_predict"
+        service_name: "bedrock"
+        model_id: "<your model id used in the ml-commons for vector search>"
+        output_path: "<your batch job output in S3>"
+        input_key: key
+        aws:
+          region: "us-east-1"
+        ml_when: /bucket == "offlinebatch"
+
+```
+`model_id` as the model id that is registered in the OpenSearch ml-commons plugin.
+`service_name` as the remote AI service platform to process then batch job.
+`output_path` as the batch job output location of the S3 Uri
+
+# Metrics
+
+### Counter
+- `mlProcessorSuccessRequests`: measures total number of requests received and processed successfully by ml-processor.
+- `mlProcessorFailedRequests`: measures total number of requests failed by ml-processor.
+- `numberOfBatchJobsCreationSucceeded`: measures total number of batch jobs successfully created (200 response status code) by OpenSearch ml-commons API.
+- `numberOfBatchJobsCreationFailed`: measures total number of batch jobs failed in creation by OpenSearch ml-commons API.
+
+## Developer Guide
+
+The integration tests for this plugin do not run as part of the Data Prepper build.
+The following command runs the integration tests:
