+++ data-prepper-plugins/ml-inference-processor/build.gradle
@@ -4,31 +4,22 @@
  */
 
 dependencies {
-    implementation project(':data-prepper-api')
     implementation project(path: ':data-prepper-plugins:common')
     implementation project(':data-prepper-plugins:aws-plugin-api')
-    implementation project(':data-prepper-plugins:failures-common')
-    implementation project(':data-prepper-plugins:parse-json-processor')
-    implementation 'io.micrometer:micrometer-core'
-    implementation 'com.fasterxml.jackson.core:jackson-core'
-    implementation 'com.fasterxml.jackson.core:jackson-databind'
-    implementation 'software.amazon.awssdk:lambda'
     implementation 'software.amazon.awssdk:sdk-core'
-    implementation 'software.amazon.awssdk:netty-nio-client'
     implementation 'software.amazon.awssdk:sts'
-    implementation 'org.hibernate.validator:hibernate-validator:8.0.1.Final'
-    implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
+    implementation 'io.micrometer:micrometer-core'
     implementation 'org.json:json'
-    implementation libs.commons.lang3
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
     implementation 'org.projectlombok:lombok:1.18.22'
+    implementation 'software.amazon.awssdk:s3'
     compileOnly 'org.projectlombok:lombok:1.18.20'
     annotationProcessor 'org.projectlombok:lombok:1.18.20'
     testCompileOnly 'org.projectlombok:lombok:1.18.20'
     testAnnotationProcessor 'org.projectlombok:lombok:1.18.20'
-    testImplementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
-    testImplementation project(':data-prepper-test-common')
-    testImplementation testLibs.slf4j.simple
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
+    testImplementation project(':data-prepper-test-event')
 }
 
 test {
@@ -60,15 +51,6 @@ task integrationTest(type: Test) {
     classpath = sourceSets.integrationTest.runtimeClasspath
 
     systemProperty 'log4j.configurationFile', 'src/test/resources/log4j2.properties'
-    
-
-    systemProperty 'tests.lambda.sink.region', System.getProperty('tests.lambda.sink.region')
-    systemProperty 'tests.lambda.sink.functionName', System.getProperty('tests.lambda.sink.functionName')
-    systemProperty 'tests.lambda.sink.sts_role_arn', System.getProperty('tests.lambda.sink.sts_role_arn')
-
-    systemProperty 'tests.lambda.processor.region', System.getProperty('tests.lambda.processor.region')
-    systemProperty 'tests.lambda.processor.functionName', System.getProperty('tests.lambda.processor.functionName')
-    systemProperty 'tests.lambda.processor.sts_role_arn', System.getProperty('tests.lambda.processor.sts_role_arn')
 
     filter {
         includeTestsMatching '*IT'
