+++ data-prepper-plugins/ml-inference-processor/src/main/java/org/opensearch/dataprepper/plugins/ml/processor/configuration/AwsAuthenticationOptions.java
