--- data-prepper-plugins/ml-processor/README.md
@@ -1,38 +1,0 @@
-
-# ml Processor
-
-This plugin enables you to send data from your Data Prepper pipeline directly to ml-commons for machine learning related activities.
-
-## Usage
-```aidl
-lambda-pipeline:
-...
-  processor:
-    - ml:
-        host: "https://search-xunzh-ml-tests-ihx7htldf7nvo2gdg25m6ehthq.us-east-1.es.amazonaws.com"
-        aws_sigv4: true
-        action_type: "batch_predict"
-        service_name: "bedrock"
-        model_id: "6ifdTZUBEBlFHJzvGSxO"
-        output_path: "s3://offlinebatch/bedrock-multisource/output-multisource/"
-        aws:
-          region: "us-east-1"
-        ml_when: /bucket == "offlinebatch"
-
-```
-`model_id` as the model id that is registered in the OpenSearch ml-commons plugin.
-`service_name` as the remote AI service platform to process then batch job.
-`output_path` as the batch job output location of the S3 Uri
-
-# Metrics
-
-### Counter
-- `mlProcessorSuccessRequests`: measures total number of requests received and processed successfully by ml-processor.
-- `mlProcessorFailedRequests`: measures total number of requests failed by ml-processor.
-- `numberOfBatchJobsCreationSucceeded`: measures total number of batch jobs successfully created (200 response status code) by OpenSearch ml-commons API.
-- `numberOfBatchJobsCreationFailed`: measures total number of batch jobs failed in creation by OpenSearch ml-commons API.
-
-## Developer Guide
-
-The integration tests for this plugin do not run as part of the Data Prepper build.
-The following command runs the integration tests:
