+++ settings.gradle
@@ -184,7 +184,7 @@ include 'data-prepper-plugins:rds-source'
 include 'data-prepper-plugins:http-source-common'
 include 'data-prepper-plugins:http-common'
 include 'data-prepper-plugins:aws-lambda'
-include 'data-prepper-plugins:ml-processor'
+include 'data-prepper-plugins:ml-inference-processor'
 //include 'data-prepper-plugins:dummy-plugin'
 include 'data-prepper-plugin-schema'
 include 'data-prepper-plugins:kinesis-source'
